packages/ADMETox/dockerfiles/requirements.txt
Django>=1.3
djangorestframework
six>=1.4.1
unicodecsv
joblib==0.14.1
numpy==1.16.5
pandas==0.24.2
djangorestframework-csv==2.1.1